refactor(app): relocate getIncompleteInstrumentCount
